Reorganize (sub)folders and comply with PEP8 (lowercase filenames)
