+++ Examples/Example_Stiffness_tensor.py
@@ -1,4 +1,4 @@
-from Elasticipy.Elasticity import StiffnessTensor
+from Elasticipy.tensors.elasticity import StiffnessTensor
 from scipy.spatial.transform import Rotation
 import matplotlib as mpl
 mpl.use('Qt5Agg')   # Ensure interactive plot
